websocket/backend-django-channels/requirements.txt
Django==4.2.14
daphne==4.1.2
channels==4.1.0
channels-redis==4.2.0
influxdb-client==1.43.0
python-dateutil==2.9.0.post0
django-cors-headers==4.4.0
redis>=4.2.0